# GlTF export performance improvements