--- a/Elements.Benchmarks/Elements.Benchmarks.csproj
+++ b/Elements.Benchmarks/Elements.Benchmarks.csproj
@@ -10,6 +10,12 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Reference Include="Csg">
+      <HintPath>../Elements/lib/Csg.dll</HintPath>
+    </Reference>
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="BenchmarkDotNet" Version="0.12.1" />
   </ItemGroup>