--- a/Elements.Benchmarks/Program.cs
+++ b/Elements.Benchmarks/Program.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.IO;
 using System.Linq;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Running;
@@ -10,22 +9,23 @@ using Elements.Serialization.glTF;
 
 namespace Elements.Benchmarks
 {
+    [MemoryDiagnoser]
     [SimpleJob(launchCount: 1, warmupCount: 1, targetCount: 3)]
-    public class Csg
+    public class CsgBenchmarks
     {
         private WideFlangeProfileFactory _profileFactory = new WideFlangeProfileFactory();
+        private Beam _beam;
+        private Csg.Solid _csg;
 
-        [Params(1, 10, 20)]
-        public int Samples { get; set; }
+        [Params(1, 10, 20, 50)]
+        public int NumberOfHoles { get; set; }
 
-        [Benchmark(Description = "Compute csg of beam.")]
-        public void CSG()
+        public CsgBenchmarks()
         {
-            var profile = _profileFactory.GetProfileByType(WideFlangeProfileType.W10x100);
-
             var line = new Line(new Vector3(0, 0, 0), new Vector3(10, 0, 5));
-            var beam = new Beam(line, profile, BuiltInMaterials.Steel);
-            for (var i = 0.0; i <= 1.0; i += 1.0 / (double)Samples)
+            var profile = Polygon.Rectangle(Units.InchesToMeters(10), Units.InchesToMeters(20));
+            _beam = new Beam(line, profile, BuiltInMaterials.Steel);
+            for (var i = 0.0; i <= 1.0; i += 1.0 / (double)NumberOfHoles)
             {
                 var t = line.TransformAt(i);
                 var lt = new Transform(t.Origin, t.ZAxis, t.XAxis.Negate());
@@ -34,11 +34,15 @@ namespace Elements.Benchmarks
                 {
                     LocalTransform = lt
                 };
-                beam.Representation.SolidOperations.Add(hole);
+                _beam.Representation.SolidOperations.Add(hole);
             }
-            var model = new Model();
-            model.AddElement(beam);
-            GltfExtensions.InitializeGlTF(model, out var buffers, false);
+            _csg = _beam.GetFinalCsgFromSolids();
+        }
+
+        [Benchmark(Description = "Tesselate CSG.")]
+        public void CsgToGraphicsBuffers()
+        {
+            _csg.Tessellate();
         }
     }
 
@@ -46,22 +50,21 @@ namespace Elements.Benchmarks
     [SimpleJob(launchCount: 1, warmupCount: 3, targetCount: 10)]
     public class HSS
     {
-        [Benchmark(Description = "Create all HSS beams.")]
-        public void CreateAllHSSBeams()
+        Model _model;
+
+        public HSS()
         {
             var x = 0.0;
             var z = 0.0;
             var hssFactory = new HSSPipeProfileFactory();
             var profiles = hssFactory.AllProfiles().ToList();
-            var model = new Model();
+            _model = new Model();
             foreach (var profile in profiles)
             {
                 var color = new Color((float)(x / 20.0), (float)(z / profiles.Count), 0.0f, 1.0f);
                 var line = new Line(new Vector3(x, 0, z), new Vector3(x, 3, z));
-                var m = new Material(Guid.NewGuid().ToString(), color, 0.1f, 0.4f);
-                model.AddElement(m, false);
-                var beam = new Beam(line, profile, m);
-                model.AddElement(beam, false);
+                var beam = new Beam(line, profile);
+                _model.AddElement(beam);
                 x += 2.0;
                 if (x > 20.0)
                 {
@@ -69,7 +72,18 @@ namespace Elements.Benchmarks
                     x = 0.0;
                 }
             }
-            model.ToGlTF();
+        }
+
+        [Benchmark(Description = "Draw all beams without merge.")]
+        public void DrawAllBeamsWithoutMerge()
+        {
+            _model.ToGlTF(false, false);
+        }
+
+        [Benchmark(Description = "Draw all beams with merge.")]
+        public void DrawAllBeamsWithMerge()
+        {
+            _model.ToGlTF(false, true);
         }
     }
 
@@ -77,7 +91,8 @@ namespace Elements.Benchmarks
     {
         public static void Main(string[] args)
         {
-            var summary = BenchmarkRunner.Run<HSS>();
+            BenchmarkRunner.Run<HSS>();
+            BenchmarkRunner.Run<CsgBenchmarks>();
         }
     }
 }