--- a/Elements/src/Geometry/Solids/Solid.cs
+++ b/Elements/src/Geometry/Solids/Solid.cs
@@ -770,6 +770,13 @@ namespace Elements.Geometry.Solids
             return loop;
         }
 
+        // TODO: This method should not be required if we have adequate lookup
+        // operations based on vertex locations and incident edges. This is 
+        // currently used in the case where we want to add a face from a polygon
+        // but in most cases where we want to do that we already know something
+        // about the shape of the existing solid and should be able to lookup 
+        // existing vertices without doing an O(n) search. Implement a better
+        // search strategy!
         private void FindOrCreateVertex(Vector3 pt, int i, Transform transform, bool mergeVerticesAndEdges, Vertex[] verts)
         {
             if (transform != null)