… Elements/src/Geometry/Vector3.cs
… Elements/src/Serialization/glTF/GltfExtensions.cs
… Elements/test/CsgTests.cs