During discussions of compiling Elements to wasm, we began to investigate sources of slowness in the gltf conversion path. There were a number of opportunities for small improvements, like reducing allocation of new polygons, and removing unnecessary creation of `Vector3` when converting from another vector type. 

- Removes the use of octree for merge vertex lookup during CSG output to graphic buffers.
- Removes use of `polygon.Transformed()` and `polygon.Reversed()` in `Solid`. Those methods both allocate new polygons which is not required as the transforms and winding can be handled while building face loops.
- Adds a couple of method to `Vector3` to allow dot and cross products to be calculated using components of an incoming vector.

Writing all HSS members to gltf before:

|                          Method |     Mean |   Error |  StdDev |      Gen 0 |     Gen 1 | Gen 2 | Allocated |
|-------------------------------- |---------:|--------:|--------:|-----------:|----------:|------:|----------:|
| 'Draw all beams without merge.' | 235.7 ms | 4.99 ms | 2.97 ms | 29000.0000 | 1000.0000 |     - |  95.18 MB |
|    'Draw all beams with merge.' | 232.3 ms | 2.69 ms | 1.60 ms | 30000.0000 | 3000.0000 |     - |  95.18 MB |

After:
|                          Method |      Mean |     Error |    StdDev |      Gen 0 |     Gen 1 |     Gen 2 | Allocated |
|-------------------------------- |----------:|----------:|----------:|-----------:|----------:|----------:|----------:|
| 'Draw all beams without merge.' | 119.72 ms | 29.135 ms | 19.271 ms | 25000.0000 | 2000.0000 | 1000.0000 |  83.01 MB |
|    'Draw all beams with merge.' |  95.55 ms |  8.368 ms |  5.535 ms | 25000.0000 | 2000.0000 | 1000.0000 |  83.01 MB |

Observations:
- The "before" benchmarks are run against master and don't have the API which allows to merge vertices, so both numbers are similar.
- The "after" benchmarks are run with and without the new API to merge vertices, which shows a nearly 20ms speed up even with the merge. My assumption here is that the reduced number of vertices after merge helps to accelerate everything downstream.
 
Creating a beam with a varying number of holes before:
|           Method | NumberOfHoles |     Mean |    Error |   StdDev |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |-------------- |---------:|---------:|---------:|--------:|------:|------:|----------:|
| 'Tesselate CSG.' |             1 | 756.0 us | 459.1 us | 25.17 us | 81.0547 |     - |     - | 250.77 KB |
| 'Tesselate CSG.' |            10 | 811.7 us | 472.5 us | 25.90 us | 81.0547 |     - |     - | 250.77 KB |
| 'Tesselate CSG.' |            20 | 774.2 us | 168.2 us |  9.22 us | 81.0547 |     - |     - | 250.77 KB |
| 'Tesselate CSG.' |            50 | 769.9 us | 399.9 us | 21.92 us | 81.0547 |     - |     - | 250.77 KB |

After:
|           Method | NumberOfHoles |     Mean |     Error |  StdDev |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |-------------- |---------:|----------:|--------:|--------:|------:|------:|----------:|
| 'Tesselate CSG.' |             1 | 143.7 us |  59.86 us | 3.28 us | 55.1758 |     - |     - | 169.75 KB |
| 'Tesselate CSG.' |            10 | 149.3 us | 169.65 us | 9.30 us | 55.1758 |     - |     - | 169.75 KB |
| 'Tesselate CSG.' |            20 | 158.4 us |  38.93 us | 2.13 us | 55.1758 |     - |     - | 169.75 KB |
| 'Tesselate CSG.' |            50 | 137.9 us | 106.55 us | 5.84 us | 55.1758 |     - |     - | 169.75 KB |

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/hypar-io/elements/656)
<!-- Reviewable:end -->